+++ components/PostHeaderAuthors.tsx
@@ -1,9 +1,7 @@
 import React, { useState } from "react";
 import Image from "next/image";
 import Link from "next/link";
-import { useRouter } from "next/router";
-import { FaLinkedin, FaTwitter, FaLink } from "react-icons/fa";
-import { sanitizeAuthorSlug } from "../utils/sanitizeAuthorSlug"; 
+import { sanitizeAuthorSlug } from "../utils/sanitizeAuthorSlug";
 
 const PostHeaderAuthors = ({ blogwriter, blogreviewer, timetoRead }) => {
   const sameAuthor =
@@ -12,7 +10,6 @@ const PostHeaderAuthors = ({ blogwriter, blogreviewer, timetoRead }) => {
 
   const [hoverStateBlogWriter, setHoverStateBlogWriter] = useState(false);
   const [hoverStateBlogReviewer, setHoverStateBlogReviewer] = useState(false);
-  const [copied, setCopied] = useState(false); 
 
   const onMouseEnterBlogWriter = () => {
     setHoverStateBlogWriter(true);
@@ -32,22 +29,6 @@ const PostHeaderAuthors = ({ blogwriter, blogreviewer, timetoRead }) => {
     setTimeout(() => {
       setHoverStateBlogReviewer(false);
     }, 400);
-  };
-  const router = useRouter();
-  const currentURL = encodeURIComponent(
-    `keploy.io/${router.basePath + router.asPath}`
-  );
-  const twitterShareUrl = `https://twitter.com/share?url=${currentURL}`;
-  const linkedinShareUrl = `https://www.linkedin.com/shareArticle?url=${currentURL}`;
-
-  const copyToClipboard = async () => {
-    try {
-      await navigator.clipboard.writeText(`https://keploy.io/blog${router.asPath}`);
-      setCopied(true);
-      setTimeout(() => setCopied(false), 2000); 
-    } catch (err) {
-      console.error("Failed to copy!", err);
-    }
   };
 
   return (
@@ -138,35 +119,6 @@ const PostHeaderAuthors = ({ blogwriter, blogreviewer, timetoRead }) => {
                 </div>
               )}
             </div>
-          )}
-        </div>
-        <div className="flex flex-row gap-5 items-center gap-3 sm:gap-5 order-2 sm:order-none mt-2 md:mb-2">
-          <p className="text-gray-500 text-sm">Share this</p>
-          <Link
-            href={twitterShareUrl}
-            target="_blank"
-            rel="noopener noreferrer"
-            className="twitter-share-button text-xl text-black transition-colors duration-300 hover:text-blue-500"
-          >
-            <FaTwitter className="icon" />
-          </Link>
-          <Link
-            href={linkedinShareUrl}
-            target="_blank"
-            rel="noopener noreferrer"
-            className="linkedin-share-button text-xl text-black transition-colors duration-300 hover:text-blue-500"
-          >
-            <FaLinkedin className="icon" />
-          </Link>
-          <button
-            onClick={copyToClipboard}
-            className="link-share-button text-xl text-black transition-colors duration-300 hover:text-blue-500 focus:outline-none"
-            aria-label="Copy URL to clipboard"
-          >
-            <FaLink className="icon" />
-          </button>
-          {copied && (
-            <span className="ml-2 text-orange-500 text-sm">Copied!</span>
           )}
         </div>
       </div>
